refactor(web,worker): improve code architecture and separation of concerns
Fix auth store layer violation by extracting auth actions into a registry
pattern (dependency inversion), split large files into focused modules
(iCal parser, useCompensations, manuscript parser, LoginPage), add
ApiClient type safety with satisfies constraints, remove deprecated
submitScorer stub from calendar client, and standardize worker error
responses to JSON format.
