+++ packages/worker/src/handlers/ocr.ts
@@ -26,10 +26,10 @@ export async function handleOcr(request: Request, env: Env, _url: URL): Promise<
 
   // Only allow POST for OCR
   if (request.method !== 'POST') {
-    return new Response('Method Not Allowed', {
+    return new Response(JSON.stringify({ error: 'Method Not Allowed' }), {
       status: 405,
       headers: {
-        'Content-Type': 'text/plain',
+        'Content-Type': 'application/json',
         Allow: 'POST',
         ...corsHeaders(origin),
         ...securityHeaders(),
