+++ packages/worker/src/utils.ts
@@ -15,3 +15,4 @@ export * from './utils/cache'
 export * from './utils/ical-validation'
 export * from './utils/session-relay'
 export * from './utils/constants'
+export * from './utils/error-response'
